Partial 64-bit update
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@1190 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/mx_bluice.c
@@ -1387,7 +1387,7 @@ mx_bluice_update_operation_status( MX_BLUICE_SERVER *bluice_server,
 		 */
 
 		snprintf( format, sizeof(format), "%%%ds",
-			sizeof(completion_status) );
+			(int) sizeof(completion_status) );
 
 		num_items = sscanf( ptr, format, completion_status );
